requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.2
django-admin-logs==1.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
